backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
websockets>=12.0
httpx==0.28.1
python-dotenv==1.2.2
huggingface_hub>=0.20.0
